Project rename and clean code
