+++ PoliDL-GUI.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30413.136
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebExAutoRecord", "WebExAutoRecord\WebExAutoRecord.csproj", "{F88550F2-17D7-029A-1719-5365FB055ADE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PoliDL-GUI", "WebExAutoRecord\PoliDL-GUI.csproj", "{F88550F2-17D7-029A-1719-5365FB055ADE}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{71AF09D3-7526-49C4-B06D-05D3563850E0}"
 EndProject
