+++ WebExAutoRecord/Forms/ProgressTracker.cs
@@ -3,7 +3,7 @@ using System;
 using System.Diagnostics;
 using System.Windows.Forms;
 
-namespace PoliDLGUI
+namespace PoliDLGUI.Forms
 {
     public partial class ProgressTracker
     {
