+++ WebExAutoRecord/Program.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using PoliDLGUI.Forms;
+using System;
 using System.Windows.Forms;
 
 namespace PoliDLGUI
